reuse URLChecker class for validateNoRedirects
