+++ src/main/java/org/opensearch/geospatial/ip2geo/dao/GeoIpDataDao.java
@@ -56,9 +56,8 @@ import org.opensearch.core.xcontent.XContentBuilder;
 import org.opensearch.geospatial.annotation.VisibleForTesting;
 import org.opensearch.geospatial.constants.IndexSetting;
 import org.opensearch.geospatial.ip2geo.common.DatasourceManifest;
-import org.opensearch.geospatial.ip2geo.common.HttpRedirectValidator;
 import org.opensearch.geospatial.ip2geo.common.Ip2GeoSettings;
-import org.opensearch.geospatial.ip2geo.common.URLDenyListChecker;
+import org.opensearch.geospatial.ip2geo.common.URLChecker;
 import org.opensearch.geospatial.shared.Constants;
 import org.opensearch.geospatial.shared.StashedThreadContext;
 import org.opensearch.index.query.QueryBuilders;
@@ -94,13 +93,13 @@ public class GeoIpDataDao {
     private final ClusterService clusterService;
     private final ClusterSettings clusterSettings;
     private final Client client;
-    private final URLDenyListChecker urlDenyListChecker;
+    private final URLChecker urlChecker;
 
-    public GeoIpDataDao(final ClusterService clusterService, final Client client, final URLDenyListChecker urlDenyListChecker) {
+    public GeoIpDataDao(final ClusterService clusterService, final Client client, final URLChecker urlChecker) {
         this.clusterService = clusterService;
         this.clusterSettings = clusterService.getClusterSettings();
         this.client = client;
-        this.urlDenyListChecker = urlDenyListChecker;
+        this.urlChecker = urlChecker;
     }
 
     /**
@@ -177,7 +176,7 @@ public class GeoIpDataDao {
         SpecialPermission.check();
         return AccessController.doPrivileged((PrivilegedAction<CSVParser>) () -> {
             try {
-                URL zipUrl = urlDenyListChecker.toUrlIfNotInDenyList(manifest.getUrl());
+                URL zipUrl = urlChecker.toUrlIfAllowed(manifest.getUrl());
                 return internalGetDatabaseReader(manifest, zipUrl.openConnection());
             } catch (IOException e) {
                 throw new OpenSearchException("failed to read geoip data from {}", manifest.getUrl(), e);
@@ -191,7 +190,7 @@ public class GeoIpDataDao {
         connection.addRequestProperty(Constants.USER_AGENT_KEY, Constants.USER_AGENT_VALUE);
 
         if (connection instanceof HttpURLConnection) {
-            HttpRedirectValidator.validateNoRedirects((HttpURLConnection) connection);
+            urlChecker.validateNoRedirects((HttpURLConnection) connection);
         }
 
         ZipInputStream zipIn = new ZipInputStream(connection.getInputStream());
