+++ src/test/java/org/opensearch/geospatial/ip2geo/dao/GeoIpDataDaoTests.java
@@ -61,8 +61,8 @@ public class GeoIpDataDaoTests extends Ip2GeoTestCase {
 
     @Before
     public void init() {
-        noOpsGeoIpDataDao = new GeoIpDataDao(clusterService, client, urlDenyListChecker);
-        verifyingGeoIpDataDao = new GeoIpDataDao(clusterService, verifyingClient, urlDenyListChecker);
+        noOpsGeoIpDataDao = new GeoIpDataDao(clusterService, client, urlChecker);
+        verifyingGeoIpDataDao = new GeoIpDataDao(clusterService, verifyingClient, urlChecker);
     }
 
     public void testCreateIndexIfNotExistsWithExistingIndex() {
@@ -131,7 +131,7 @@ public class GeoIpDataDaoTests extends Ip2GeoTestCase {
         assertArrayEquals(expectedHeader, parser.iterator().next().values());
         String[] expectedValues = { "1.0.0.0/24", "Australia" };
         assertArrayEquals(expectedValues, parser.iterator().next().values());
-        verify(urlDenyListChecker).toUrlIfNotInDenyList(manifest.getUrl());
+        verify(urlChecker).toUrlIfAllowed(manifest.getUrl());
     }
 
     public void testGetDatabaseReaderNoFile() throws Exception {
@@ -146,7 +146,7 @@ public class GeoIpDataDaoTests extends Ip2GeoTestCase {
         );
         Exception exception = expectThrows(IllegalArgumentException.class, () -> noOpsGeoIpDataDao.getDatabaseReader(manifest));
         assertTrue(exception.getMessage().contains("does not exist"));
-        verify(urlDenyListChecker).toUrlIfNotInDenyList(manifest.getUrl());
+        verify(urlChecker).toUrlIfAllowed(manifest.getUrl());
     }
 
     @SneakyThrows
