+++ src/test/java/org/opensearch/geospatial/plugin/GeospatialPluginTests.java
@@ -39,7 +39,7 @@ import org.opensearch.geospatial.ip2geo.action.RestUpdateDatasourceHandler;
 import org.opensearch.geospatial.ip2geo.common.Ip2GeoExecutor;
 import org.opensearch.geospatial.ip2geo.common.Ip2GeoLockService;
 import org.opensearch.geospatial.ip2geo.common.Ip2GeoSettings;
-import org.opensearch.geospatial.ip2geo.common.URLDenyListChecker;
+import org.opensearch.geospatial.ip2geo.common.URLChecker;
 import org.opensearch.geospatial.ip2geo.dao.DatasourceDao;
 import org.opensearch.geospatial.ip2geo.dao.GeoIpDataDao;
 import org.opensearch.geospatial.ip2geo.dao.Ip2GeoCachedDao;
@@ -64,13 +64,13 @@ import org.opensearch.watcher.ResourceWatcherService;
 
 public class GeospatialPluginTests extends OpenSearchTestCase {
     private final ClusterSettings clusterSettings = new ClusterSettings(Settings.EMPTY, new HashSet(Ip2GeoSettings.settings()));
-    private final URLDenyListChecker urlDenyListChecker = new URLDenyListChecker(clusterSettings);
+    private final URLChecker urlChecker = new URLChecker(clusterSettings);
     private final List<RestHandler> SUPPORTED_REST_HANDLERS = List.of(
         new RestUploadGeoJSONAction(),
         new RestUploadStatsAction(),
-        new RestPutDatasourceHandler(clusterSettings, urlDenyListChecker),
+        new RestPutDatasourceHandler(clusterSettings, urlChecker),
         new RestGetDatasourceHandler(),
-        new RestUpdateDatasourceHandler(urlDenyListChecker),
+        new RestUpdateDatasourceHandler(urlChecker),
         new RestDeleteDatasourceHandler()
     );
